requirements.txt
annotated-types==0.6.0
astroid==3.1.0
blinker==1.7.0
click==8.1.7
dill==0.3.8
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
packaging==23.2
platformdirs==4.2.0
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
pylint==3.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.27
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.10.0
Werkzeug==3.0.1
